refactor: move back to utils
